bacports-wireless: Rename cfg80211 -> cy_cfg80211
The backports-wireless (brcmfmac) is an external module
package that builds cfg80211.ko among other modules.
Rename that to cy_cfg80211.ko to avoid a conflict with
the "in-tree" driver, which has the same name.

The lkm_loader script loads the appropriate driver
(cfg80211 or cy_cfg80211) based on whether bcmdhd or
brcmfmac drivers are required.

Bug 2511019

Change-Id: I8574c217bda9d19e6dfddba9e4c62919a5aee449
Signed-off-by: Abhiroop Kaginalkar <akaginalkar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2117971
GVS: Gerrit_Virtual_Submit
Reviewed-by: Om Prakash Singh <omp@nvidia.com>
Tested-by: Om Prakash Singh <omp@nvidia.com>
Reviewed-by: Mohan Thadikamalla <mohant@nvidia.com>
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>